CbTarragonDriver: improve DiodeFault detection (refs EVEREST-707)
In case of CP state changes from A -> C we currently are not able to
detect the diode fault correctly since EvseManager enables the PWM
later that we report CP state C. At this moment, we are already in
CP state C and thus do not detect a change.
For this reason we have to also check for possible diode fault when
the duty cycle change occurs.

Signed-off-by: Fabian Hartung <fabian.hartung@chargebyte.com>
Signed-off-by: Martin Lukas <martin.lukas@chargebyte.com>
Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>