Fix unexpected diode fault when switching between PWM states
The issue was caused by a race condition between the thread that sets
the PWM value and the thread that observes the CP state. Now the
CP observation thread has to wait for the CP state reading
until the new PWM value is set.

Signed-off-by: Fabian Hartung <fabian.hartung@chargebyte.com>

Disable CP observation via disable_cp_observation() method

Changed log level from info to debug level for the CP observation status
message to avoid flooding of the log.

Signed-off-by: Fabian Hartung <fabian.hartung@chargebyte.com>