Fix deletion of lock error
In case of a not properly connected plug, a lock error is raised.
This error would only be cleared when a plug lock is correctly closed again.
This can lead to the assumption that the EVSE is broken.

Since we don't react to CP state changes directly, but to commands coming from EvseManager,
the lock error will now be cleared when the plug lock is in state "unlock".

This "unlock" state is essential, because when the plug lock is not in an unlocked state,
the EVSE can't be used properly.

Signed-off-by: Martin Lukas <martin.lukas@chargebyte.com>