CbSystem: Add Secure Firmware Update of OCPP 2.0.1 (refs EVEREST-1110)
Now the charging station sends a FirmwareStatusNotificationRequest with
status "Installed" after rebooting of the board for the secured and common
firmware update mechanism. Before rebooting the board, the charging
station sends a FirmwareStatusNotificationRequest with status "InstallRebooting".
Added a flag in "/var/lib/everest/.ocpp_fw_installed" to detect that the
firmware has been successfully installed after rebooting the board.

Signed-off-by: Fabian Hartung <fabian.hartung@chargebyte.com>