CbChargeSOMDriver: add initial implementation
This adds initial support for the upcoming Charge SOM platform.

It is based on the preliminary communication protocol of the safety controller.
Supported so far:
 - CP generation and evaluation
 - PP evaluation

The code base is not yet very clean, due to the overall preliminary
nature of this material. Also the documentation, example config etc.
is still missing.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>
Reviewed-by: Mohannad Oraby <mohannad.oraby@chargebyte.com>
Reviewed-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>
Reviewed-by: Stefan Wahren <stefan.wahren@chargebyte.com>