CbChargeSOMDriver: Fix invalid CP state in case of short circuit between CP and PE (refs EVEREST-1184)
In case CP state changes from C to E (shorted CP to PE) it is possible
that the positive side is shortly in state D. The state D transition
should be ignored in this case. This check must be done before checking
for the ventilation error to avoid false positives.

Signed-off-by: Fabian Hartung <fabian.hartung@chargebyte.com>