CbTarragonDriver: refactoring of contactor support
This splits the contactor functionality into three dedicated classes which
implement the 'mutual', 'serial' wirings for 3ph-1ph setups, and the general
'simple' wiring for usual setups without phase count switching.

It also moves a lot of functionality down to the relais class, so that
upper layers a more light-weight.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>