CbTarragonDriver: fix RCM corner case while power on throttling is active
In case the RCM reports an error while the power on throttling is active,
e.g. we are waiting to close the contactor, then we reported a contactor
error. But this is not the case here, we just skipped the real contactor
closing, so we have to report success in this case to prevent the
upstream error generation.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>