CbSystem: set system time via SDBus instead of command line utility (#19)
* CbSystem: provide libsystemd via pkg-config

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

* CbSystem: implement setSystemTime() and use it

This replaces use of /bin/date, which cannot handle the RFC3339 date, if
implemented by busybox.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

* CbSystem: use exception handling in setSystemTime()

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

* CbSystem: report result from sd_bus_error_is_set()

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

* CbSystem: update time on every call, but only with significant differences

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

* CbSystem: convert parsed time point to system clock

Everest::Date's from_rfc3339() uses date::utc_clock, which takes leap
seconds into account. This gave an offset to the system time of currently
28 seconds.

Convert to system time to set the correct date.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

* CbSystem: increase the minimum time devation for setting the time

Apparently, setting the system time sometimes takes more than one seconds.
To avoid setting the time every time (e.g. with every OCPP Heartbeat),
increase the minimm deviation to two seconds.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

* CbSystem: use explicit type for SDBus function

This ensures that the `sd_bus_call_method()` API is used correctly.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

* CbSystem: close SDBus in error paths

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

* CbSystem: use `this` consistently

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

* CbSystem: use std::chrono more consistently

This computes differences in the std::chrono domain, and allows using
literals.

Move the constant into the implementation, like the other constants.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

* CbSystem: reduce significant time difference

500 ms deviation seems more realistic.

The setting of the time itself has some delay itself, but that is only in
the 10s of milliseconds.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

* CbSystem: make system time difference configurable

Also increase the default back to 1000 ms, as a compromise for slower
connections (e.g. communication via mobile network).

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

* CbSystem: convert config variables to snake case

The variables, as copied from the upstream System module, broke with the
convention to use snake_case. Fix this, and also convert the "number" types
to "integer", as they are integral anyway (counters), or cast to an integer
in the code (time in seconds).

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

* CbSystem: switch to sd_bus_flush_close_unrefp cleanup

This simplifies closing.

Also evaluate errors more extensively, and ignore EALREADY (system clock is
already in sync).

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

---------

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>