CbChargeSOMDriver: rewrite to adapt for newer UART protocol
This change is a complete rewrite of the existing module.
It requires the newer safety controller firmware (>= 0.2.x) and
implements the newer UART protocol which provides a more
higher-level access/view of the charging process.

This is necessary since the safety controller already takes care
of various low-level functionalities, combined with the complete
safety control mechanisms.

Part of this safety implementation is also that PT1000 sensors
are monitored. The values are also provided as dedicated interface.

Another point to mention is that for the low-level UART access
a dedicated library is used now.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>