Fix various cppcheck warnings (#33)
* CbSystem: do not return pointer to local variable

Not sure whether a copy is created on implicit conversion to its return type
(fs::path), but cppcheck complained that the returned pointer goes out of
scope.

Create an explicit conversion, which should in turn create a copy.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

* CbSystem: initialize variables

This may be redundant with the operations in init(), but cppcheck complains.
(The fact that init() is launched first before other operations is implicit
by the framework.)

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

* CbTarragonDriver: fix style

Merge subsequent identical if() clauses.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

---------

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>