Factor out the actuator handling from relay class
This moves the actuator stuff out of the current class by using the
new CbActuator class. The rough idea is that all the modelling still
applies, but the actual actuator switching (and thus relay switching)
is done using the new class which guarantees that all GPIOs are
switched in sync.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>