CbTarragonDriver: clear MREC23ProximityFault on EV disconnect (closes #35)
We set a PP fault in two cases: either a real signal level occurred, or
the PP was disconnected during charging. When the EV is then disconected,
we can and should clear the error flag. Either the error is really gone,
or in case of a persistent error, it is immediately raised again after
the next measurement round.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>